Arrange test code into more packages
