+++ app/src/test/java/com/carkzis/ananke/ui/AnankeAppStateTest.kt
@@ -1,4 +1,4 @@
-package com.carkzis.ananke
+package com.carkzis.ananke.ui
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
@@ -10,8 +10,6 @@ import androidx.navigation.compose.composable
 import androidx.navigation.createGraph
 import androidx.navigation.testing.TestNavHostController
 import com.carkzis.ananke.navigation.AnankeDestination
-import com.carkzis.ananke.ui.AnankeAppState
-import com.carkzis.ananke.ui.rememberAnankeAppState
 import org.junit.Assert.*
 import org.junit.Before
 import org.junit.Rule
